requirements.txt
discord.py
openai
requests
python-dotenv
pymongo[svr,tls]
colorlog
